+++ CHANGELOG.md
@@ -15,6 +15,8 @@ All notable changes to this project will be documented in this file. This projec
 ### Added
 
 - `check-munki-pkgsinfo` now produces an error if `uninstall_method` is set to `uninstall_script` but no uninstall script is present in the pkginfo.
+- `check-munki-pkgsinfo` now checks for deprecated pkginfo keys.
+- `check-munki-pkgsinfo` now includes checks for many possible pkginfo key typos, not just `minimum_os_version` and `maximum_os_version`. Suggestions welcome if you think of more.
 - `check-munkiadmin-scripts` now checks whether scripts are named correctly, not just executable.
 
 ## [1.17.0] - 2024-12-22